Runs on command line inputs
Created a matA, matB and matC for every possible size, and split the mygemm function based on those sizes. Updated main to only instantiate the matrices we are using, and only start threads for those matrices to prevent artificial inflation of the execution time.

You'll need to change the execution time formula to the one you found that works and run it again to make sure it is consistent with the changes. But we should be able to write a script now that runs all the different combinations of THREAD_COUNT and SIZE.

The command line format is ./mpl THREAD_COUNT SIZE just for clarity.